一部変数の数値型を変更しました
